+++ sekoia_valhalla_integration_modules/__init__.py
@@ -1,10 +1,10 @@
 from sekoia_automation.module import Module
 
-from sekoia_valhalla_integration_modules.actions.delete_catalog_rules import (
-    DeleteCatalogRules,
-)
 from sekoia_valhalla_integration_modules.models import (
     SekoiaValhallaIntegrationModuleConfiguration,
+)
+from sekoia_valhalla_integration_modules.triggers.delete_catalog_rules import (
+    DeleteCatalogRules,
 )
 from sekoia_valhalla_integration_modules.triggers.sync_sigma_intelligence_center import (
     SyncSigmaIntelligenceCenter,